refactor: 필터 옵션 및 정렬 기능을 위한 상수와 타입 정의 추가
