+++ src/UI/pages/filter/components/SortingFilter.ts
@@ -1,16 +1,12 @@
 import RestaurantFilter from '../../../components/filter/RestaurantFilter';
-import { SortOption } from '../../../../Domain/types/FilterOption';
+import { SortOption } from '../../../../Domain/types/FilterOptionType';
+import { SORTING_LIST } from '../../../../Domain/constants/FilterOption';
 
 class SortingFilter {
   private filter: RestaurantFilter<SortOption>;
 
   constructor(onChange?: (sortBy: SortOption) => void) {
-    const sortingOptions = [
-      { value: 'name' as SortOption, text: '이름순' },
-      { value: 'distance' as SortOption, text: '거리순' },
-    ];
-
-    this.filter = new RestaurantFilter<SortOption>(sortingOptions, onChange);
+    this.filter = new RestaurantFilter<SortOption>(SORTING_LIST, onChange);
     this.filter.getElement().id = 'sorting-filter';
     this.filter.getElement().name = 'sorting';
   }
